+++ SaanSoft.cqrs.sln
@@ -15,7 +15,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SaanSoft.Tests.Cqrs.Common"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SaanSoft.Cqrs.Decorator.Store.MongoDB", "src\SaanSoft.Cqrs.Decorator.Store.MongoDB\SaanSoft.Cqrs.Decorator.Store.MongoDB.csproj", "{73DF9405-CA2A-4A08-AEF2-1549471DD0AB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB", "test\SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB\SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB.csproj", "{2F2E1D89-634E-4CC4-AED6-179DAA952A93}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB.GuidIds", "test\SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB.GuidIds\SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB.GuidIds.csproj", "{2F2E1D89-634E-4CC4-AED6-179DAA952A93}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_solution", "_solution", "{8962FA21-773D-4F0E-81CB-AF1D7AE5B9A7}"
 	ProjectSection(SolutionItems) = preProject
@@ -36,6 +36,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SaanSoft.Tests.Cqrs.Decorator", "test\SaanSoft.Tests.Cqrs.Decorator\SaanSoft.Tests.Cqrs.Decorator.csproj", "{241D036D-F677-4F05-94F9-BCE8FBE671DE}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SaanSoft.Cqrs.GuidIds", "src\SaanSoft.Cqrs.GuidIds\SaanSoft.Cqrs.GuidIds.csproj", "{AF159B96-B46C-4998-B748-EC86F1F45C8C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SaanSoft.Cqrs.Decorator.Store.MongoDB.GuidIds", "src\SaanSoft.Cqrs.Decorator.Store.MongoDB.GuidIds\SaanSoft.Cqrs.Decorator.Store.MongoDB.GuidIds.csproj", "{EAA2420F-0957-45D9-ADDD-E1DE2152AB60}"
 EndProject
 Global
 				GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -79,6 +81,10 @@ Global
 					{AF159B96-B46C-4998-B748-EC86F1F45C8C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 					{AF159B96-B46C-4998-B748-EC86F1F45C8C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 					{AF159B96-B46C-4998-B748-EC86F1F45C8C}.Release|Any CPU.Build.0 = Release|Any CPU
+					{EAA2420F-0957-45D9-ADDD-E1DE2152AB60}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+					{EAA2420F-0957-45D9-ADDD-E1DE2152AB60}.Debug|Any CPU.Build.0 = Debug|Any CPU
+					{EAA2420F-0957-45D9-ADDD-E1DE2152AB60}.Release|Any CPU.ActiveCfg = Release|Any CPU
+					{EAA2420F-0957-45D9-ADDD-E1DE2152AB60}.Release|Any CPU.Build.0 = Release|Any CPU
 				EndGlobalSection
 				GlobalSection(NestedProjects) = preSolution
 								{14FB4E51-D7DF-4045-BF3D-9635FFE3E309} = {446FC180-C478-471E-A0A4-DB7B76C0B674}
