+++ test/SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB.GuidIds/GlobalUsings.cs
@@ -4,6 +4,6 @@ global using Microsoft.Extensions.DependencyInjection;
 global using MongoDB.Driver;
 global using SaanSoft.Cqrs.Decorator.Store.Models;
 global using SaanSoft.Cqrs.Decorator.Store.MongoDB;
-global using SaanSoft.Cqrs.GuidIds.Messages;
+global using SaanSoft.Cqrs.Decorator.Store.Utilities;
 global using SaanSoft.Cqrs.Messages;
 global using SaanSoft.Tests.Cqrs.Common.TestModels;
