+++ test/SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB.GuidIds/Repositories/CommandHandlerRepositoryTests.cs
@@ -1,4 +1,4 @@
-namespace SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB.Repositories;
+namespace SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB.GuidIds.Repositories;
 
 public class CommandHandlerRepositoryTests : BaseHandlerRepositoryTests<MyCommand, CommandRepository, BaseCommand<Guid>>
 {
