+++ test/SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB.GuidIds/Repositories/QueryHandlerRepositoryTests.cs
@@ -1,4 +1,4 @@
-namespace SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB.Repositories;
+namespace SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB.GuidIds.Repositories;
 
 public class QueryHandlerRepositoryTests : BaseHandlerRepositoryTests<MyQuery, QueryRepository, Query<Guid>>
 {
